Upgrade to go 1.24 and fix lint errors (#681)
* Remove deprecated linters

* Use new range in benchmark loops

* Add timestamp overflow check for validator registrations

* Define gasLimit as a uint64

* Replace "golang.org/x/exp/slices" with stdlib "slices"

* Use require.JSONEq

* Add a bunch of //nolint:gosec comments for timestamp conversions

* Disable deprecated tenv linter

* Define new errTimestampOverflow var

* Bump golang to v1.24

* Run go mod tidy

* Replace context.Background() with t.Context() in tests

+++ internal/investigations/validator-registration-signature-check/main.go
@@ -16,7 +16,7 @@ import (
 )
 
 var (
-	gasLimit     = 30000000
+	gasLimit     = uint64(30000000)
 	feeRecipient = "0xdb65fEd33dc262Fe09D9a2Ba8F80b329BA25f941"
 	timestamp    = int64(1606824043)
 )
@@ -39,7 +39,7 @@ func main() {
 
 	// Fill in validator registration details
 	validatorRegistration := builderApiV1.ValidatorRegistration{ //nolint:exhaustruct
-		GasLimit:  uint64(gasLimit),
+		GasLimit:  gasLimit,
 		Timestamp: time.Unix(timestamp, 0),
 	}
